Term version read and window hide added
Now the term version is read and checked if it fits to the library.
Added a function to hide the term window. This is useful for
integrations to prevent the user from manually operating the GUI when
running a measurement remotely.
Added example of how to measure DC with Remote at different currents and
then measure EIS.
Thales/Term version 5.8.6 is required.